Extract verifier_observation.rs from turn.rs (parent #680)
Seventh vertical-slice extraction: move the task-contract verifier
observation hooks out of turn.rs into a new
src/agent/loop_run/verifier_observation.rs sibling module.

Migrated (all pub(super) free fns over &mut Agent):

- record_task_contract_verifier_invocation — capture most-recent
  verifier command + exit code in session state for downstream
  consumers (case_photon_bridge, eval_log). Uses
  `session::feedback::redact_verifier_command_for_storage` as SSOT
  redactor before storing.
- observe_task_contract_verifier_exit_zero — record a
  verifier_exit_zero completion-evidence entry from a non-bound run.
- observe_task_contract_verifier_exit_zero_bound — Issue #651 PR-001
  structured-runner variant. Records a bound
  `bound_test_artifacts_count` so
  `TaskContract::evaluate_with_owned_test_artifacts` can satisfy
  `test_execution_required = true`.

No facade re-export (DR3-001). Caller updates:
- verifier_orchestration.rs (×4): `agent.X(...)` →
  `super::verifier_observation::X(agent, ...)`

`cargo fix --tests` pruned 1 newly-unused import from turn.rs's prelude.

turn.rs: 5,801 -> 5,737 (-64 LOC)
Cumulative from 39,489: -33,752 LOC (-85.5%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
